Move MissingTaskDependencyDetector to own class
